# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Storyteller's Stage – API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/fairy-tales:
    get:
      tags: [Admin]
      operationId: listAdminFairyTales
      summary: List all fairy tales with translation status
      responses:
        "200":
          description: All fairy tales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminFairyTaleSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Admin]
      operationId: createAdminFairyTale
      summary: Create a new fairy tale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FairyTaleBaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/fairy-tales/{id}:
    get:
      tags: [Admin]
      operationId: getAdminFairyTale
      summary: Get a fairy tale with all its translations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Fairy tale with all translations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Admin]
      operationId: updateAdminFairyTale
      summary: Update fairy tale base metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FairyTaleBaseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: deleteAdminFairyTale
      summary: Delete a fairy tale and all its translations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/fairy-tales/{id}/translations/{locale}:
    put:
      tags: [Admin]
      operationId: upsertFairyTaleTranslation
      summary: Create or update a translation for a fairy tale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FairyTaleTranslationInput"
      responses:
        "200":
          description: Translation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/fairy-tales/{id}/translations/{locale}/publish:
    post:
      tags: [Admin]
      operationId: publishFairyTaleTranslation
      summary: Publish a translation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Translation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/fairy-tales/{id}/translations/{locale}/unpublish:
    post:
      tags: [Admin]
      operationId: unpublishFairyTaleTranslation
      summary: Unpublish a translation (back to draft)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFairyTaleDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Translation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── Admin – Plays ────────────────────────────────────────────────────────
...
  /admin/about/{locale}:
    get:
      tags: [Admin]
      operationId: getAdminAbout
      summary: Get the About bio for a locale (admin)
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: About content for this locale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAbout"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [Admin]
      operationId: saveAdminAbout
      summary: Create or update the About bio for a locale
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AboutInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAbout"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TranslationStatus:
      type: object
      required: [locale, status]
      properties:
        locale:
          type: string
        status:
          type: string
          description: "draft | published | archived | missing"
        publishedAt:
          type: [string, "null"]
          format: date-time

    # ── Fairy Tales – Admin ───────────────────────────────────────────────────
    AdminFairyTaleSummary:
      type: object
      required: [id, slug, createdAt, updatedAt, translations]
      properties:
        id:
          type: number
        slug:
          type: string
        ageRecommendation:
          type: [string, "null"]
        estimatedReadingTime:
          type: [number, "null"]
        themes:
          type: [array, "null"]
          items:
            type: string
        coverImagePath:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        translations:
          type: array
          items:
            $ref: "#/components/schemas/TranslationStatus"
    AdminFairyTaleTranslation:
      type: object
      required: [id, fairyTaleId, locale, status, title, createdAt, updatedAt]
      properties:
        id:
          type: number
        fairyTaleId:
          type: number
        locale:
          type: string
        status:
          type: string
        title:
          type: string
        blurb:
          type: [string, "null"]
        body:
          type: [string, "null"]
        seoTitle:
          type: [string, "null"]
        seoDescription:
          type: [string, "null"]
        publishedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminFairyTaleDetail:
      type: object
      required: [id, slug, createdAt, updatedAt, translations]
      properties:
        id:
          type: number
        slug:
          type: string
        ageRecommendation:
          type: [string, "null"]
        estimatedReadingTime:
          type: [number, "null"]
        themes:
          type: [array, "null"]
          items:
            type: string
        coverImagePath:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        translations:
          type: array
          items:
            $ref: "#/components/schemas/AdminFairyTaleTranslation"
    FairyTaleBaseInput:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          minLength: 1
        ageRecommendation:
          type: [string, "null"]
        estimatedReadingTime:
          type: [number, "null"]
        themes:
          type: [array, "null"]
          items:
            type: string
        coverImagePath:
          type: [string, "null"]
    FairyTaleTranslationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        blurb:
          type: [string, "null"]
        body:
          type: [string, "null"]
        seoTitle:
          type: [string, "null"]
        seoDescription:
          type: [string, "null"]

    # ── Fairy Tales – Public ──────────────────────────────────────────────────
...
    AdminAbout:
      type: object
      required: [locale, updatedAt]
      properties:
        locale:
          type: string
        body:
          type: [string, "null"]
        authorPhotoPath:
          type: [string, "null"]
        updatedAt:
          type: string
          format: date-time
    AboutInput:
      type: object
      properties:
        body:
          type: [string, "null"]
        authorPhotoPath:
          type: [string, "null"]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...